Modify struct fields' contextstr
