+++ tests/struct_field.rs
@@ -107,3 +107,32 @@ fn finds_struct_field_in_constructor() {
     }"#;
     assert_eq!(get_definition(src, None).matchstr, "id");
 }
+
+#[test]
+fn struct_field_scalar_primitive_types() {
+    let src = "
+    struct Foo<'a> {
+        reference: &'a u8,
+        array: [u8; 5],
+        slice: &'a [u8],
+    }
+
+    fn foo(x: Foo) {
+        x.~
+    }
+    ";
+
+    let completions = get_all_completions(src, None);
+    assert_eq!(completions.len(), 3);
+
+    for completion in completions {
+        let expected = match completion.matchstr.as_ref() {
+            "reference" => "reference: &'a u8",
+            "array" => "array: [u8; 5]",
+            "slice" => "slice: &'a [u8]",
+            _ => panic!("unexpected match from Foo struct ({})", completion.matchstr),
+        };
+
+        assert_eq!(completion.contextstr, expected);
+    }
+}
